requirements.txt
addict==2.4.0
ansi2html==1.8.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.6
cmake==3.25.0
comm==0.1.3
ConfigArgParse==1.7
contourpy==1.1.0
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.9.0
Flask==2.2.5
fonttools==4.41.1
fvcore==0.1.5.post20221221
gdown==4.7.1
idna==3.4
imageio==2.31.1
iopath==0.1.10
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
lit==15.0.7
lpips==0.1.4
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
open3d==0.17.0
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.9.1
plotly==5.15.0
portalocker==2.7.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
Pygments==2.15.1
pyparsing==3.0.9
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@v0.7.4
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.0
requests==2.28.1
retrying==1.3.4
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trimesh==3.22.5
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wrapt==1.15.0
yacs==0.1.8
-f https://build.rerun.io/commit/29ab0e3/wheels 
rerun-sdk==0.9.0a4
